Prefer the tab character for indentation.
When we create snippets, we could indent 2 spaces, or 4, or using
tabs, etc. However, that impresses our opinion on other developers.

If we use the tab character TextMate will interpret it as if the
developer had pressed the tab key on their keyboard, and will use
the developers custom indentation settings.